+++ src/hla/__init__.py
@@ -37,7 +37,7 @@ Modules include:
         
 """
 
-__version__ = "0.2.0"
+__version__ = "0.2.2"
 
 
 #import os, sys, re